+++ packages/config/src/configs/index.ts
@@ -33,6 +33,10 @@ import { nodleConfig } from './nodle';
 import { originTrailAlphanetConfig } from './originTrailAlphanet';
 import { parallelConfig } from './parallel';
 import { parallelHeikoConfig } from './parallelHeiko';
+import { peaqConfig } from './peaq';
+import { peaqAlphanetConfig } from './peaqAlphanet';
+import { peaqEvmConfig } from './peaqEvm';
+import { peaqEvmAlphanetConfig } from './peaqEvmAlphanet';
 import { pendulumConfig } from './pendulum';
 import { pendulumAlphanetConfig } from './pendulumAlphanet';
 import { phalaConfig } from './phala';
@@ -80,6 +84,10 @@ export const chainsConfigList: ChainConfig[] = [
   originTrailAlphanetConfig,
   parallelConfig,
   parallelHeikoConfig,
+  peaqConfig,
+  peaqEvmConfig,
+  peaqAlphanetConfig,
+  peaqEvmAlphanetConfig,
   pendulumConfig,
   pendulumAlphanetConfig,
   phalaConfig,