+++ packages/config/src/configs/peaqAlphanet.ts
@@ -1,0 +1,43 @@
+import {
+  AssetMinBuilder,
+  BalanceBuilder,
+  ExtrinsicBuilder,
+} from '@moonbeam-network/xcm-builder';
+import { agng, dev, ftmwh } from '../assets';
+import { moonbaseAlpha, peaqAlphanet } from '../chains';
+import { AssetConfig } from '../types/AssetConfig';
+import { ChainConfig } from '../types/ChainConfig';
+
+export const peaqAlphanetConfig = new ChainConfig({
+  assets: [
+    new AssetConfig({
+      asset: dev,
+      balance: BalanceBuilder().substrate().assets().account(),
+      destination: moonbaseAlpha,
+      destinationFee: {
+        amount: 0.01,
+        asset: dev,
+        balance: BalanceBuilder().substrate().assets().account(),
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transfer(),
+      min: AssetMinBuilder().assets().asset(),
+    }),
+    new AssetConfig({
+      asset: ftmwh,
+      balance: BalanceBuilder().substrate().assets().account(),
+      destination: moonbaseAlpha,
+      destinationFee: {
+        amount: 0.04,
+        asset: dev,
+        balance: BalanceBuilder().substrate().assets().account(),
+      },
+      extrinsic: ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
+      fee: {
+        asset: agng,
+        balance: BalanceBuilder().substrate().system().account(),
+      },
+      min: AssetMinBuilder().assets().asset(),
+    }),
+  ],
+  chain: peaqAlphanet,
+});