+++ packages/sdk/src/contract/contracts/Xtokens/Xtokens.ts
@@ -1,6 +1,7 @@
 import { ContractConfig } from '@moonbeam-network/xcm-builder';
 import { Contract, TransactionResponse } from 'ethers';
 import {
+  Address,
   GetContractReturnType,
   PublicClient,
   WriteContractReturnType,
@@ -16,7 +17,10 @@ import abi from './XtokensABI.json';
 type XtokensContract = GetContractReturnType<typeof abi, PublicClient>;
 
 export class Xtokens implements TransferContractInterface {
-  readonly address = '0x0000000000000000000000000000000000000804';
+  readonly defaultMoonChainAddress =
+    '0x0000000000000000000000000000000000000804';
+
+  readonly address: Address;
 
   readonly #config: ContractConfig;
 
@@ -24,10 +28,12 @@ export class Xtokens implements TransferContractInterface {
 
   readonly #signer: EvmSigner;
 
-  constructor(config: ContractConfig, signer: EvmSigner) {
+  constructor(config: ContractConfig, signer: EvmSigner, address?: Address) {
     this.#config = config;
 
     this.#signer = signer;
+
+    this.address = address ?? this.defaultMoonChainAddress;
 
     this.#contract = isEthersSigner(signer)
       ? new Contract(this.address, abi, signer)
@@ -56,7 +62,6 @@ export class Xtokens implements TransferContractInterface {
     return contract[this.#config.func].estimateGas(...this.#config.args);
   }
 
-  // eslint-disable-next-line class-methods-use-this
   async getViemContractEstimatedGas(
     contract: XtokensContract,
   ): Promise<bigint> {
@@ -64,12 +69,8 @@ export class Xtokens implements TransferContractInterface {
       return 0n;
     }
 
-    // Temporary fix to Module(ModuleError { index: 51, error: [0, 0, 0, 0], message: None })
-    // To be reverted next week when Moonbeam is in RT2900
-    return 100000n;
-
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    // return contract.estimateGas[this.#config.func](this.#config.args as any);
+    return contract.estimateGas[this.#config.func](this.#config.args as any);
   }
 
   async getFee(amount: bigint): Promise<bigint> {
@@ -89,7 +90,6 @@ export class Xtokens implements TransferContractInterface {
     } catch (error) {
       // eslint-disable-next-line no-console
       console.error(error);
-
       throw new Error(
         "Can't get a fee. Make sure that you have enough balance!",
       );