+++ packages/types/src/chain/parachain/EvmParachain.ts
@@ -1,4 +1,4 @@
-import { defineChain } from 'viem';
+import { Address, defineChain } from 'viem';
 import { Chain } from 'viem/chains';
 import { ChainType } from '../Chain.interfaces';
 import { Parachain, ParachainConstructorParams } from './Parachain';
@@ -8,12 +8,18 @@ export interface EvmParachainConstructorParams
   id: number;
   rpc: string;
   nativeCurrency: NativeCurrency;
+  isEvmSigner?: boolean;
+  contracts?: Contracts;
 }
 
 type NativeCurrency = {
   decimals: number;
   name: string;
   symbol: string;
+};
+
+type Contracts = {
+  Xtokens?: Address;
 };
 
 export class EvmParachain extends Parachain {
@@ -23,17 +29,25 @@ export class EvmParachain extends Parachain {
 
   readonly nativeCurrency: NativeCurrency;
 
+  readonly isEvmSigner: boolean;
+
+  readonly contracts?: Contracts;
+
   constructor({
     id,
     rpc,
     nativeCurrency,
+    isEvmSigner = false,
+    contracts,
     ...others
   }: EvmParachainConstructorParams) {
     super({ type: ChainType.EvmParachain, ...others });
 
+    this.contracts = contracts;
     this.id = id;
     this.rpc = rpc;
     this.nativeCurrency = nativeCurrency;
+    this.isEvmSigner = isEvmSigner;
   }
 
   getViemChain(): Chain {